Add unit test for FileScanner, move test to correct package
