+++ src/test/resources/nl/aerius/sldgenerator/plugin/custom_draw.sld
