+++ src/test/resources/nl/aerius/sldgenerator/plugin/image.sld
