+++ src/test/resources/nl/aerius/sldgenerator/plugin/simple.sld.json
