+++ src/test/resources/nl/aerius/sldgenerator/plugin/simple_with_zoomlevels.sld.json
