+++ src/test/resources/nl/aerius/sldgenerator/plugin/zoom_levels_default.json
